Update load and store instuctions for rv64
Also this fixes a bug in backstepping FSD and a tautologically true test.

+++ src/rars/riscv/instructions/LB.java
@@ -35,7 +35,7 @@ public class LB extends Load {
         super("lb t1, -100(t2)", "Set t1 to sign-extended 8-bit value from effective memory byte address", "000");
     }
 
-    public int load(int address) throws AddressErrorException {
+    public long load(int address) throws AddressErrorException {
         return (Globals.memory.getByte(address) << 24) >> 24; // Shifting sign extends
     }
 }
